chore(e2e): kustomize util refactoring
